% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_plots.R
\docType{class}
\name{CategoricalPlot-class}
\alias{CategoricalPlot-class}
\alias{categorical_plot}
\title{CategoricalPlot}
\description{
Handle Configuration and output of a categorical variable plot
}
\section{Methods}{

\describe{
\item{\code{plot(..., x_axis, y_axis, group_by)}}{Prepare ggplot output for the plot

            @description
            Create a line or bar plot from objects data and fields metadata.
            @param x_axis column name from data to be plotted on the x axis
            @param y_axis column name from data to be plotted on the y axis
            @param group_by vector of category column names
            @param type determies plot type; either 'line' or 'bar'}

\item{\code{set_to_bar(...)}}{Set type of plot returned to a barchart}

\item{\code{set_to_line(...)}}{Set type of plot returned to a barchart}
}}

