% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{class}
\name{GeneralPlot-class}
\alias{GeneralPlot-class}
\alias{general_plot}
\title{Superclass to define plot object initialization}
\arguments{
\item{fields}{metadata for the data columns}

\item{data}{data.frame for the plot data}

\item{x_axis}{column name from data to be plotted on the x axis}

\item{y_axis}{column name from data to be plotted on the y axis}

\item{group_by}{vector of category column names}

\item{type}{determies plot type; either 'line' or 'bar'}
}
\description{
Handle Configuration and output of a categorical variable plot
}

