% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_plots.R
\docType{class}
\name{NumericPlot-class}
\alias{NumericPlot-class}
\alias{numeric_plot}
\title{NumericPlot}
\arguments{
\item{x_axis}{column name from data to be plotted on the x axis}

\item{y_axis}{column name from data to be plotted on the y axis}

\item{group_by}{vector of category column names"}
}
\description{
Handle Configuration and output of a numeric variable plot
}
\section{Methods}{

\describe{
\item{\code{plot(...)}}{Create a numerical plot from data and settings.}
}}

