% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_plots.R
\docType{class}
\name{CategoricalPlot-class}
\alias{CategoricalPlot-class}
\alias{categorical_plot}
\title{CategoricalPlot}
\arguments{
\item{fields}{metadata for the data columns}

\item{data}{data.frame for the plot data}

\item{x_axis}{column name from data to be plotted on the x axis}

\item{y_axis}{column name from data to be plotted on the y axis}

\item{dimension_metadata}{list of selected category column names as keys and
selected row values as list values.}

\item{group_axis}{vector with column names used as group argument in ggplot}

\item{type}{determies plot type; either 'line' or 'bar'}
}
\description{
Handle Configuration and output of a categorical variable plot
}
\section{Methods}{

\describe{
\item{\code{plot(...)}}{Prepare ggplot output from data and config.}

\item{\code{set_to_bar(...)}}{Set type of plot returned to a barchart}

\item{\code{set_to_line(...)}}{Set type of plot returned to a barchart}
}}

