% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_plots.R
\docType{class}
\name{NumericPlot-class}
\alias{NumericPlot-class}
\alias{numeric_plot}
\title{NumericPlot}
\arguments{
\item{fields}{metadata for the data columns}

\item{data}{data.frame for the plot data}

\item{x_axis}{column name from data to be plotted on the x axis}

\item{y_axis}{column name from data to be plotted on the y axis}

\item{dimension_metadata}{vector of category column names}

\item{type}{determies plot type; either 'line' or 'bar'}
}
\description{
Handle Configuration and output of a numeric variable plot
}
\section{Methods}{

\describe{
\item{\code{plot(...)}}{Create a numerical plot from data and settings.}

\item{\code{set_to_boxplot(...)}}{Set type of plot returned to a barchart}

\item{\code{set_to_line(...)}}{Set type of plot returned to a barchart}

\item{\code{set_y_scale_limits(..., y_scale_limits)}}{Set y scale upper and lower limit.}
}}

