% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_plots.R
\docType{class}
\name{NumericPlot-class}
\alias{NumericPlot-class}
\alias{numeric_plot}
\title{NumericPlot}
\description{
Handle Configuration and output of a numeric variable plot        df
}
\section{Methods}{

\describe{
\item{\code{plot(..., x_axis = "year", y_axis = "number", group_by = c())}}{Put together a numerical plot

            @description
            Create a line plot from data.
            @param x column name from data to be plotted on the x axis
            @param y column name from data to be plotted on the y axis
            @param group vector of category column names}
}}

